+++ packages/design-system/src/index.ts
@@ -30,7 +30,14 @@ import HeaderBar from './components/HeaderBar';
 import { Icon } from './components/Icon';
 import { IconsProvider } from './components/IconsProvider';
 import InlineEditing from './components/InlineEditing';
-import InlineMessage from './components/InlineMessage';
+import {
+	InlineMessage,
+	InlineMessageInformation,
+	InlineMessageSuccess,
+	InlineMessageWarning,
+	InlineMessageDestructive,
+	InlineMessageBeta,
+} from './components/InlineMessage';
 import Layout from './components/Layout';
 import Link from './components/Link';
 import { LinkAsButton } from './components/LinkAsButton';
@@ -43,6 +50,7 @@ import Skeleton, {
 	SkeletonHeading,
 	SkeletonButton,
 	SkeletonButtonIcon,
+	SkeletonInput,
 } from './components/Skeleton';
 import {
 	StatusCanceled,
@@ -104,6 +112,11 @@ export {
 	IconsProvider,
 	InlineEditing,
 	InlineMessage,
+	InlineMessageInformation,
+	InlineMessageSuccess,
+	InlineMessageWarning,
+	InlineMessageDestructive,
+	InlineMessageBeta,
 	Layout,
 	Link,
 	LinkAsButton,
@@ -115,6 +128,7 @@ export {
 	SkeletonButton,
 	SkeletonButtonIcon,
 	SkeletonHeading,
+	SkeletonInput,
 	SkeletonParagraph,
 	Status,
 	StatusCanceled,